EcommerceDjango/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.13
django-cors-headers==4.6.0
python-dotenv==1.0.1
stripe==9.10.0
google-generativeai>=0.8.0
